Move Python files in a new speculos/ directory
This prepares work to package Speculos as a Python package.

Most imports of other modules were modified to relative imports.

`speculos.py` has been re-created in the root directory, to ensure
backward compatibility with existing instructions to launch Speculos.
