+++ speculos/api/resources/apdu.schema
