+++ speculos/api/resources/button.schema
