+++ speculos/api/resources/finger.schema
