+++ speculos/api/swagger.yaml
