+++ speculos/api/swagger/swagger-ui-standalone-preset.js
