+++ speculos/mcu/automation.py
