+++ speculos/mcu/automation_server.py
