+++ speculos/mcu/icon.png
