+++ speculos/mcu/nanox_ocr.py
