+++ speculos/mcu/resources/automation.schema
