+++ speculos/mcu/seproxyhal.py
